+++ frontend/src/constants.js
@@ -1,1 +1,1 @@
-export const AR_FRONTEND_VERSION = '0.7.2';
+export const AR_FRONTEND_VERSION = '0.8.0';